v.0.1.5

+++ pkg/repository/tables.go
@@ -1,7 +1,7 @@
 package repository
 
 type Users struct {
-	ID       uint64  `json:"user_id"`
+	ID       uint32  `json:"user_id"`
 	Username string  `json:"username"`
 	Name     string  `json:"name"`
 	Surname  string  `json:"surname"`
@@ -10,7 +10,7 @@ type Users struct {
 }
 
 type Messages struct {
-	ID     uint64 `json:"message_id"`
+	ID     uint32 `json:"message_id"`
 	Text   string `json:"text"`
 	Time   string `json:"sending_time"`
 	Status int    `json:"message_status"`
@@ -18,12 +18,12 @@ type Messages struct {
 }
 
 type Chats struct {
-	ID      uint64   `json:"chat_id"`
+	ID      uint32   `json:"chat_id"`
 	UsersID []uint64 `json:"users_id"`
 }
 
 type Transactions struct {
-	ID          uint64  `json:"transaction_id"`
+	ID          uint32  `json:"transaction_id"`
 	Amount      float64 `json:"amount"`
 	SenderID    uint64  `json:"sender_id"`
 	RecipientID uint64  `json:"recipient_id"`
